setting up permission for roles and added page module

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Gate;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Spatie\Permission\Models\Permission;
+use Symfony\Component\HttpFoundation\Response;
 
 class PermissionController extends Controller
 {
@@ -15,14 +17,16 @@ class PermissionController extends Controller
      */
     public function index(Request $request)
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $keyword = $request->get('search');
         $perPage = 25;
 
         if (!empty($keyword)) {
             $permission = Permission::where('name', 'LIKE', "%$keyword%")
-                ->latest()->paginate($perPage);
+                ->orderBy('name','asc')->paginate($perPage);
         } else {
-            $permission = Permission::latest()->paginate($perPage);
+            $permission = Permission::orderBy('name','asc')->paginate($perPage);
         }
 
         return view('admin.permission.index', compact('permission'));
@@ -35,6 +39,8 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_add'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.permission.create');
     }
 
@@ -47,6 +53,7 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('permission_add'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         
         $requestData = $request->all();
         
@@ -78,6 +85,8 @@ class PermissionController extends Controller
      */
     public function edit($id)
     {
+        abort_if(Gate::denies('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permission = Permission::findOrFail($id);
 
         return view('admin.permission.edit', compact('permission'));
@@ -93,6 +102,7 @@ class PermissionController extends Controller
      */
     public function update(Request $request, $id)
     {
+        abort_if(Gate::denies('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         
         $requestData = $request->all();
         
@@ -111,6 +121,9 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
+
+        abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         Permission::destroy($id);
 
         return redirect('admin/user/permission')->with('flash_message', 'Permission deleted!');
